Simply usage example.
I simplified the usage example. There were `async` and `await` keywords that were unnecessary and misleading because the `Reactor` constructor is synchronous. I also removed the global error handling. If the customer wants to implement error handling, it would normally be done on specific API calls and not at a global level. I don't think we need to go into error handling in this usage example.